url_screenshot/server/requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
by==0.0.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.8
distro==1.9.0
et-xmlfile==1.1.0
fastapi==0.110.1
groq==0.4.2
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
lxml==5.1.0
macholib==1.16.3
Markdown==3.5.2
numpy==1.26.4
openai==1.13.3
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pydantic==2.6.3
pydantic_core==2.16.3
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyMySQL==1.1.0
pynput==1.7.6
pyobjc-core==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-Quartz==10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
screeninfo==0.8.1
selenium==4.18.1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
